pest-ai/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.10.6
pydantic_core==2.27.2
httpx==0.27.2
pydantic-ai==0.0.21
openai==1.60.2
supabase==2.12.0
loguru==0.7.2
python-multipart==0.0.9
dspy
google-genai==1.2.0
google-generativeai==0.8.4
google-ai-generativelanguage==0.6.15